Extract Config interface to allow alternate implementations for storage
