+++ src/main/java/com/xero/api/XeroClient.java
@@ -99,11 +99,7 @@ public class XeroClient {
 	
 	protected static final Pattern MESSAGE_PATTERN = Pattern.compile("<Message>(.*)</Message>");
 	protected final ObjectFactory objFactory = new ObjectFactory();
-	
-	public XeroClient() {
-		config = Config.getInstance();
-	}
-	
+
 	public XeroClient(Config config) {
 		this.config = config;
 	}
